[mypyc] Refactor: more fallback primitive ops to a new module
