+++ mypyc/irbuild/for_helpers.py
@@ -16,7 +16,7 @@ from mypyc.ir.rtypes import (
 )
 from mypyc.primitives.int_ops import unsafe_short_add
 from mypyc.primitives.list_ops import new_list_op, list_append_op, list_get_item_unsafe_op
-from mypyc.primitives.misc_ops import iter_op, next_op
+from mypyc.primitives.generic_ops import iter_op, next_op
 from mypyc.primitives.exc_ops import no_err_occurred_op
 from mypyc.irbuild.builder import IRBuilder
